# api/openapi.yaml
openapi: 3.0.3
info:
  title: Toskar Control Plane API
  version: 1.4.0
  description: >-
    Local AI control plane API for Toskar, its OpenAI-compatible routes, and its MCP server.
    On the default loopback bind no key is needed. When the daemon listens on any other address,
    every /api/v1, /v1, and /mcp route needs Authorization: Bearer YOUR_API_KEY; /about and /source do not.
    Errors are {"error": {"code", "message", "details"}}. docs/api.md describes the behavior of each route.
servers:
  - url: http://127.0.0.1:7331
security:
  - {}
  - bearerAuth: []
paths:
...
  /api/v1/me/devices/{id}:
    delete:
      summary: Disconnect one of your devices (#206)
      operationId: disconnectMyDevice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Disconnected; it needs a new code to connect again
        "404":
          description: No device of yours has that id
...
  /api/v1/devices/pairing:
    post:
      summary: Show a 6-digit code a phone connects with, replacing any other
      operationId: startDevicePairing
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                enable_lan:
                  type: boolean
                  description: Turn on local network access first, which a phone needs
      responses:
        "201":
          description: The code, when it expires, this computer's address, and whether a phone can reach it yet
    get:
      summary: Whether a phone has connected with the code shown
      operationId: getDevicePairing
      responses:
        "200":
          description: The state (waiting, connected, expired, or cancelled) and the phone's key record, without the code
    delete:
      summary: Stop showing the code
      operationId: cancelDevicePairing
      responses:
        "204":
          description: Stopped
  /api/v1/devices/pair:
    post:
      summary: A phone exchanges the code for its own key (no key needed; local network only)
      operationId: pairDevice
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  example: "123456"
                device_name:
                  type: string
                  example: Sam's iPhone
      responses:
        "201":
          description: The phone's key (api_key, shown once) and its record, with kind device
        "401":
          description: PAIRING_WRONG_CODE
        "403":
          description: PAIRING_NOT_LOCAL
        "410":
          description: PAIRING_NOT_STARTED or PAIRING_EXPIRED
        "429":
          description: PAIRING_THROTTLED
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: An API key from the API Access page, or TOSKAR_API_KEY
...